observability routes
